Side by side left and right panels. iframes.
content.js extracts the html of the current page.
Clears the current page.
Creates a div (app) inserts it into an empty body tag
and renders the left and right panels into this div.

RightPanel is passed the page html to place into an iframe

public/css/root.css contains most of the css.
	There is some junk in this file. ignore it:
	.left, .right, .box

Some work remains to be done for this phase:
	the way we recreate the page holding the two panels.
		We should have a component that does this.
		see: RecreatePage.js
	The way we extract the current page's html.
		There must be a cleaner way.

 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   public/css/root.css
	modified:   scripts/start.js
	modified:   src/App.css
	modified:   src/App.js
	new file:   src/LeftPanel.css
	new file:   src/LeftPanel.js
	new file:   src/RecreatePage.js
	modified:   src/RightPanel.js
	new file:   src/Sandbox.js
	modified:   src/content.js